feat: добавлен игнор сториса в тестах, если он апрувнут
